+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 
+## 2025-05-30 11:01:06 - Commit cd4ec98
+
+- Auto-deploy: 2025-05-30 11:01:01
+
+### Fichiers modifiés:
+- CHANGELOG.md
+- backend/.env.example
+- deploy/deploy-status.json
+- frontend/.env.example
+
+
 ## 2025-05-30 10:58:34 - Commit 388a1b5
 
 - Auto-deploy: 2025-05-30 10:58:27